requirements.txt
affine==2.3.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
decorator==4.4.2
Fiona==1.8.4
GDAL==2.3.3
gdal2tiles==0.1.9
geojson==2.5.0
geomet==0.3.0
geopandas==0.8.2
html2text==2020.1.16
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
munch==2.5.0
networkx==2.5
numpy==1.20.0
osmtogeojson==0.0.2
pandas==1.2.1
Pillow==8.1.0
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
rasterio==1.1.3
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.0
sentinelsat==0.14
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
tifffile==2021.2.1
tqdm==4.56.0
urllib3==1.26.3
